// src/backend/executiveBots.d.ts
export declare const executiveBots: {
  ceo: string[];
  coo: string[];
  cfo: string[];
  cio: string[];
  cto: string[];
  cmo: string[];
  chro: string[];
  vp_sales: string[];
  vp_product: string[];
  vp_operations: string[];
  vp_marketing: string[];
  digital_marketing: string[];
  seo_specialist: string[];
  ppc_specialist: string[];
  data_scientist: string[];
  ml_engineer: string[];
  software_engineer: string[];
  cloud_architect: string[];
  blockchain_developer: string[];
  strategy: string[];
  business_analyst: string[];
  growth_hacker: string[];
  innovation_manager: string[];
  brand_strategist: string[];
  venture_capitalist: string[];
  investment_banker: string[];
  risk_manager: string[];
  financial_advisor: string[];
  supply_chain_manager: string[];
  manufacturing_engineer: string[];
  transportation_manager: string[];
  aviation_manager: string[];
  legal_counsel: string[];
  compliance_officer: string[];
  ux_designer: string[];
  architect: string[];
  civil_engineer: string[];
  structural_engineer: string[];
  fashion_designer: string[];
  sustainability_officer: string[];
  biotech_scientist: string[];
  genetic_engineer: string[];
  medical_doctor: string[];
  neurosurgeon: string[];
  ai_ethics_researcher: string[];
  cybersecurity_specialist: string[];
  content_creator: string[];
  public_relations: string[];
  journalist: string[];
  diplomat: string[];
  music_producer: string[];
  film_director: string[];
  actor: string[];
  food_beverage_director: string[];
  hospitality_manager: string[];
  customer_success: string[];
  call_center_manager: string[];
  sales_business_development: string[];
  cold_calling: string[];
  lead_qualification: string[];
  retail_manager: string[];
  ecommerce_director: string[];
  luxury_brand_manager: string[];
  sports_coach: string[];
  sports_agent: string[];
  education_consultant: string[];
  university_president: string[];
  real_estate_broker: string[];
  construction_manager: string[];
  vr_ar_developer: string[];
  game_designer: string[];
  esports_coach: string[];
  event_planner: string[];
  operations_efficiency: string[];
  political_campaign_manager: string[];
  strategic_partnerships: string[];
  renewable_energy_engineer: string[];
  electric_vehicle_engineer: string[];
  aerospace_engineer: string[];
  marketing: string[];
};